Fix theme explorer and make more responsive
