--- docs/static/theme-explorer.css
@@ -1,66 +1,0 @@
-.te-navbar {
-  /* navbar */
-  display: -ms-flexbox;
-  display: flex;
-  -ms-flex-wrap: wrap;
-  flex-wrap: wrap;
-  -ms-flex-align: center;
-  align-items: center;
-  -ms-flex-pack: justify;
-  justify-content: space-between;
-  padding: 0.5rem 1rem;
-
-  /* bg-dark */
-  background-color: #343a40 !important;
-
-  /* sticky-top */
-  position: -webkit-sticky;
-  position: sticky;
-  top: 0;
-  z-index: 1020;
-
-  /* .navbar-expand */
-  -ms-flex-flow: row nowrap;
-  flex-flow: row nowrap;
-  -ms-flex-pack: start;
-  justify-content: flex-start;
-
-  margin: 0;
-  font-size: 1rem;
-  font-weight: 400;
-  line-height: 1.5;
-  text-align: left;
-}
-
-.te-navbar-nav {
-  display: flex;
-  padding-left: 0;
-  margin-bottom: 0;
-  list-style: none;
-
-  -ms-flex-direction: row;
-  flex-direction: row;
-}
-
-.te-nav-link {
-  /* navbar-expand nav-link */
-  padding-right: 0.5rem;
-  padding-left: 0.5rem;
-
-  /* navbar-dark nav-link */
-  color: rgba(255, 255, 255, 0.5);
-
-  /* nav-link */
-  display: block;
-  padding: 0.5rem 1rem;
-
-  font-family: -apple-system, BlinkMacSystemFont, 'Segoe UI', Roboto,
-    'Helvetica Neue', Arial, 'Noto Sans', sans-serif, 'Apple Color Emoji',
-    'Segoe UI Emoji', 'Segoe UI Symbol', 'Noto Color Emoji';
-}
-
-.te-nav-link:hover,
-.te-nav-link:focus {
-  text-decoration: none;
-  color: rgba(255, 255, 255, 0.75);
-}
