+++ docs/templates/macros/theme-explorer-navbar.html
@@ -1,5 +1,5 @@
 {% macro theme_explorer_navbar() -%}
-<nav class="te-navbar">
+<nav class="navbar navbar-expand navbar-dark sticky-top bg-dark">
   <span
     class="mr-2 bg-white"
     style="padding:2px;border-radius:4px;margin-right: 0.5rem;"
@@ -11,12 +11,9 @@
       alt="dbc logo thumbnail"
     />
   </span>
-  <ul class="te-navbar-nav mr-auto">
+  <ul class="navbar-nav mr-auto">
     <li>
-      <a class="te-nav-link" href="/docs/themes">Back to themes</a>
-    </li>
-    <li>
-      <a href="#sourceCode" class="te-nav-link">Jump to source</a>
+      <a class="nav-link" href="/docs/themes">Back to themes</a>
     </li>
   </ul>
 </nav>
